requirements.txt
absl-py==2.1.0
aiosqlite==0.20.0
amqp==5.1.1
ansicon==1.89.0
APScheduler==3.10.4
asgiref==3.8.1
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.2.0
autobahn==23.1.2
Automat==22.10.0
backcall==0.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
blessed==1.20.0
blinker==1.8.2
build==1.2.1
cachetools==5.3.3
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
channels==3.0.4
channels-redis==4.0.0
chardet==5.2.0
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
constantly==15.1.0
contourpy==1.2.1
cron_descriptor==1.2.35
cryptography==40.0.2
cssselect==1.2.0
cycler==0.12.1
daphne==3.0.2
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
Django==4.2.9
django-appconf==1.0.5
django-celery-beat==2.5.0
django-celery-results==2.2.0
django-compressor==4.3.1
django-crispy-forms==2.0
django-environ==0.10.0
django-guardian==2.4.0
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
django-redis==5.2.0
django-timezone-field==6.1.0
docx2pdf==0.1.8
editdistance==0.8.1
editor==1.6.6
efficientnet==1.0.0
essential-generators==1.0
executing==1.2.0
filelock==3.14.0
Flask==3.0.3
Flask-Mail==0.10.0
flatbuffers==24.3.25
fonttools==4.51.0
gast==0.5.4
google==3.0.0
google-api-core==2.19.0
google-auth==2.29.0
google-cloud==0.34.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-oauth==1.0.1
google-pasta==0.2.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
groupdocs-conversion-cloud==24.4
grpcio==1.64.0
gunicorn==20.1.0
h11==0.14.0
h5py==3.11.0
hyperlink==21.0.0
idna==3.4
imageio==2.34.1
imgaug==0.4.0
imutils==0.5.4
incremental==22.10.0
inquirer==3.2.4
ipython==8.14.0
itemadapter==0.9.0
itemloaders==1.2.0
itsdangerous==2.2.0
jedi==0.18.2
jellyfish==1.0.3
Jinja2==3.1.4
jinxed==1.2.1
jmespath==1.0.1
jsonfield==3.1.0
keras==3.3.3
Keras-Applications==1.0.8
keras-ocr @ git+https://github.com/faustomorales/keras-ocr.git@e8d34a46f07d50158e1d86d6c617e99bfe99e2f8
kiwisolver==1.4.5
kombu==5.3.5
lazy_loader==0.4
libclang==18.1.1
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.3.2
msgpack==1.0.5
namex==0.0.8
networkx==3.3
numpy==1.26.4
opcua==0.98.13
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
outcome==1.3.0.post0
packaging==24.0
parsel==1.9.1
parso==0.8.3
path==16.14.0
pathlib==1.0.1
pexpect==4.8.0
phonenumbers==8.13.11
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==4.2.2
plum-dispatch==1.7.4
prompt-toolkit==3.0.38
Protego==0.3.1
proto-plus==1.23.0
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyaes==1.6.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
pycparser==2.21
PyDispatcher==2.0.7
Pygments==2.15.1
pyOpenSSL==23.1.1
pyparsing==3.1.2
pyproject_hooks==1.1.0
PySocks==1.7.1
pytesseract==0.3.10
python-crontab==2.7.1
python-dateutil==2.8.2
python-docx==1.1.2
pytz==2023.3
pyuseragents==1.0.5
pywin32==306
queuelib==1.7.0
rcssmin==1.1.1
readchar==4.1.0
redis==4.4.0
reportlab==4.2.0
requests==2.29.0
requests-file==2.0.0
rich==13.7.1
rjsmin==1.2.1
rsa==4.9
runs==1.2.2
safeIO==1.2
scikit-image==0.23.2
scipy==1.13.0
Scrapy==2.11.2
selenium==4.21.0
service-identity==21.1.0
setuptools==69.5.1
shapely==2.0.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
Spire.Doc==12.4.0
Spire.Pdf==10.5.2
sqlparse==0.4.4
stack-data==0.6.2
supervisor==4.2.5
Telethon==1.35.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.4.0
tifffile==2024.5.10
tldextract==5.1.2
tqdm==4.66.4
traitlets==5.9.0
translatepy==2.3
trio==0.25.1
trio-websocket==0.11.1
Twisted==22.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.11.0
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.15
validators==0.28.1
vine==5.1.0
virtualenv==20.26.2
w3lib==2.1.2
Wand==0.6.13
wcwidth==0.2.6
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
wsproto==1.2.0
xmod==1.8.1
zope.interface==6.0
